service/ai-security/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
scikit-learn==1.4.1.post1
testcontainers[redis]==4.4.0
pandas==2.2.1
numpy==1.26.4
kafka-python==2.0.2
redis==5.0.3
pytest==8.1.1
pytest-mock==3.12.0
